version2 had the goal to speed up the confusion kernel with respect to the performance obtained in version 1 of the application. to achieve so, the casting of the operands used inside the confusion kernel was turned from double to float. a little more detailed summary on the changes of this version is found in an about text inside the appropriate version2Profiling folder. after the aplied changes, as observed by the profiling files, version 2 contains the code that leads to the best performance of the confusion kernel seen so far. see profiling files and confusion kernel of this version for more details

+++ src/confusion_kernel.cu
@@ -26,10 +26,10 @@ __global__ void inverseConfusionKernel ( const unsigned char *input, unsigned ch
         int x = x_base + i ;
         if ( x < width ){  // maybe can remove this if with padding in th future ... or other strategies... i mean this
             
-            double theta = 2* (double)y / ((double) width );
-            double sinVal = sinpi( theta ) ; //use sin function for double precision, sinf for single precision
-            int y1 = ((static_cast<int>(y - x + sc * sinVal) % width) + width) % width  ;  // doing this because % operator in cuda cant do right operation for negative operands
-            int x1 = ((static_cast<int>( x - sc * sinVal) % width ) + width ) % width  ;
+            float theta = 2* (float)y / ((float) width );
+            float sinVal = sinpi( theta ) ; //use sin function for double precision, sinf for single precision
+            int y1 = ((static_cast<int>( (float)y - (float)x + (float)sc * sinVal) % width) + width) % width  ;  // doing this because % operator in cuda cant do right operation for negative operands
+            int x1 = ((static_cast<int>( (float)x - (float)sc * sinVal) % width ) + width ) % width  ;
 
             //printf("pixel in location (%d,%d), now goes to location (%d,) \n", x,y, x1 ) ;
             for(int c = 0; c<NUM_CHANNELS; c++){
@@ -96,9 +96,9 @@ __global__ void confusionKernel( const unsigned char *input, unsigned char * out
         if ( x < width ){  // maybe can remove this if with padding in th future ... or other strategies... i mean this
             
             int y1 = (y + x) % width;
-            double theta = 2.0 * (double)y1 / ( (double) width ) ;
-            double sinVal = sinpi(theta);
-            double op = (double)x + ( ( ( double ) sc) * sinVal ) ;
+            float theta = 2.0 * (float)y1 / ( (float) width ) ;
+            float sinVal = sinpi(theta);
+            float op = (float)x + ( ( ( float ) sc) * sinVal ) ;
             int x1 = ( ( (static_cast<int>( op ) % (width) ) + (width) ) % (width) ) ; 
             for(int c = 0; c < NUM_CHANNELS; c ++){
                 output[NUM_CHANNELS*( y1* width + x1 ) + c ] = input[ NUM_CHANNELS * ( y * width +x) + c ];
